Remove `Mode`.
`BuildKind` and `Mode` provide overlapping functionality, but `Mode` is
less general, so this patch removes it. As a result, the `BuildKind`
selection logic is now entirely within `main.rs`.

This patch also changes how `build_kinds` and `run_kinds` are specified,
moving the logic entirely into `main.rs`. As part of this, it
changes the default `build_kinds` selection for the `profile` subcommand
from `Debug` to `All`, and the default `run_kinds` selection from
`Check` to `All` -- in both cases to match the `bench_local` subcommand
-- and fixes the documentation accordingly.
